WIP: Refactor to Diamond standard
* Refactor/merged CannonFacetTest.js into CannonTest.js, reusing the individual tests, but using a different before hook to test as a standalone contract and as a diamond facet
